+++ cpp/Platform.Hashing/Platform.Hashing.Combine.h
@@ -17,7 +17,7 @@ namespace Platform::Hashing
         hashAccumulator = Internal::crc32(reinterpret_cast<const std::uint8_t*>(value), length * sizeof(T), hashAccumulator);
     }
 
-    void Combine(std::uint32_t& hashAccumulator, auto&& value)
+    void Combine(std::size_t& hashAccumulator, auto&& value)
     {
         Combine(hashAccumulator, &value, 1);
     }
